% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_params.R
\name{mk_coxph_params}
\alias{mk_coxph_params}
\title{Create Cox Proportional Hazards Model Parameters}
\usage{
mk_coxph_params(data, id_col, time_col, event_col, covariates_col, ...)
}
\arguments{
\item{data}{A data frame containing the ids, covariates and survival data.}

\item{id_col}{A string specifying the column name for the subject IDs.}

\item{time_col}{A string specifying the column name for the survival time.}

\item{event_col}{A string specifying the column name for the event indicator
(1 if event occurred, 0 otherwise).}

\item{covariates_col}{A vector of strings specifying the column names for the
covariates.}

\item{...}{Additional arguments to be passed to the \code{\link[survival:coxph]{survival::coxph()}}
function except for the \code{formula} and \code{data} arguments.}
}
\value{
A \code{model_params} object.
}
\description{
This function prepares the parameters required to fit a Cox proportional
hazards model using the \code{\link[survival:coxph]{survival::coxph()}} function.
}
\examples{
df <- data.frame(
  sample_id = 1:10,
  time = rexp(10),
  event = sample(0:1, 10, replace = TRUE),
  age = rnorm(10, 50, 10),
  treatment = sample(0:1, 10, replace = TRUE)
)
mk_coxph_params(
  data = df,
  id_col = "sample_id",
  time_col = "time",
  event_col = "event",
  covariates_col = c("age", "treatment")
)
}
