% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_params.R
\name{mk_model_params}
\alias{mk_model_params}
\title{Create Model Parameters}
\usage{
mk_model_params(
  data,
  id_col,
  data_col,
  fun,
  fun_params,
  header,
  gt = "genotype"
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{id_col}{A string specifying the column name in \code{data} that contains
the IDs.}

\item{data_col}{A string specifying the column name in \code{data} that contains
the data.}

\item{fun}{A function to be used for modeling.}

\item{fun_params}{A list of parameters to be passed to \code{fun}.}

\item{header}{A character vector specifying the header names.}

\item{gt}{A string specifying the genotype. Default is "genotype".}
}
\value{
A list with class "model_params" containing the following elements:
\describe{
\item{id}{The IDs extracted from \code{data} based on \code{id_col}.}
\item{gt}{The genotype string.}
\item{data}{The data extracted from \code{data} based on \code{data_col}.}
\item{fun}{The function to be used for modeling.}
\item{fun_params}{The list of parameters to be passed to \code{fun}.}
\item{header}{The header names.}
}
}
\description{
This function creates a \code{model_params} object.
}
\details{
todo!
}
\examples{
data <- data.frame(id = letters[1:5], x = rnorm(5), y = rnorm(5))
fun <- function(formula, data, ...) {
  res <- extract_lm(formula, data, ...)
  return(data.frame(coef, se, p))
}
fun_params <- list(formula = y ~ genotype + x, data = NULL)
header <- c("coef", "se", "p")
mk_model_params(data, "id", c("x", "y"), fun, fun_params, header)

}
